+++ src/document/document.cpp
@@ -845,7 +845,7 @@ ERROR CMDExpunge(void)
    {
       parasol::Log log;
       log.msg("Freeing %d internally allocated fonts.", glTotalFonts);
-      for (WORD i=0; i < glTotalFonts; i++) acFree(glFonts[i].Font);
+      for (LONG i=0; i < glTotalFonts; i++) acFree(glFonts[i].Font);
    }
 
    if (exsbuffer)         { FreeResource(exsbuffer); exsbuffer = NULL; }
